backend/requirements.txt
flask
flask-cors
deep-translator
google-generativeai
python-dotenv
qrcode
pillow
httpx
